# storage/app/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'relay-api API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://relay-api.blumilk.local.env'
...
paths:
  /api/register:
    post:
      summary: ''
      operationId: postApiRegister
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: zbailey@example.net
                password:
                  type: string
                  description: ''
                  example: architecto
              required:
                - name
                - email
                - password
      security: []
  /api/login:
    post:
      summary: ''
      operationId: postApiLogin
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: gbailey@example.net
                password:
                  type: string
                  description: ''
                  example: '|]|{+-'
              required:
                - email
                - password
      security: []
  /api/auth/google:
    post:
      summary: ''
      operationId: postApiAuthGoogle
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider_token:
                  type: string
                  description: ''
                  example: architecto
              required:
                - provider_token
      security: []
  /api/auth/2fa/verify:
    post:
      summary: ''
      operationId: postApiAuth2faVerify
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                intermediate_token:
                  type: string
                  description: ''
                  example: architecto
                code:
                  type: string
                  description: ''
                  example: architecto
              required:
                - intermediate_token
                - code
      security: []
  /api/password/email:
    post:
      summary: ''
      operationId: postApiPasswordEmail
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: gbailey@example.net
              required:
                - email
      security: []
  /api/password/reset:
    post:
      summary: ''
      operationId: postApiPasswordReset
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: ''
                  example: architecto
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: zbailey@example.net
                password:
                  type: string
                  description: 'Must be at least 8 characters.'
                  example: '-0pBNvYgxw'
              required:
                - token
                - email
                - password
      security: []
  /api/public/mobile/password/email:
    post:
      summary: ''
      operationId: postApiPublicMobilePasswordEmail
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. The <code>email</code> of an existing record in the users table.'
                  example: gbailey@example.net
              required:
                - email
      security: []
  /api/public/mobile/password/reset:
    post:
      summary: ''
      operationId: postApiPublicMobilePasswordReset
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. The <code>email</code> of an existing record in the users table.'
                  example: gbailey@example.net
                pin:
                  type: string
                  description: 'Must be 6 digits.'
                  example: '569775'
                password:
                  type: string
                  description: ''
                  example: architecto
              required:
                - email
                - pin
                - password
      security: []
  '/api/devices/{device_uuid}':
    get:
      summary: ''
      operationId: getApiDevicesDevice_uuid
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 2
                  uuid: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
                  name: 'Device 0678'
                  type: A/C
                  model: Model-5040
                  brand: 'Boyle, Zulauf and Glover'
                  serial_number: SN-23782372
                  location: "348 Kaylah Cove Suite 376\nLake Opalshire, DE 77841-5195"
                  installation_date: '2022-11-09T00:00:00.000000Z'
                  notes: 'Culpa earum neque similique quia qui vitae consequatur tenetur. Cum eligendi rerum autem molestias ullam et. Occaecati nihil placeat libero.'
                  created_at: '2026-05-16T11:06:47.000000Z'
                  updated_at: '2026-05-16T11:06:47.000000Z'
                  instruction_url: null
                properties:
                  id:
                    type: integer
                    example: 2
                  uuid:
                    type: string
                    example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
                  name:
                    type: string
                    example: 'Device 0678'
                  type:
                    type: string
                    example: A/C
                  model:
                    type: string
                    example: Model-5040
                  brand:
                    type: string
                    example: 'Boyle, Zulauf and Glover'
                  serial_number:
                    type: string
                    example: SN-23782372
                  location:
                    type: string
                    example: "348 Kaylah Cove Suite 376\nLake Opalshire, DE 77841-5195"
                  installation_date:
                    type: string
                    example: '2022-11-09T00:00:00.000000Z'
                  notes:
                    type: string
                    example: 'Culpa earum neque similique quia qui vitae consequatur tenetur. Cum eligendi rerum autem molestias ullam et. Occaecati nihil placeat libero.'
                  created_at:
                    type: string
                    example: '2026-05-16T11:06:47.000000Z'
                  updated_at:
                    type: string
                    example: '2026-05-16T11:06:47.000000Z'
                  instruction_url:
                    type: string
                    example: null
                    nullable: true
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  '/api/devices/{device_uuid}/events':
    get:
      summary: ''
      operationId: getApiDevicesDevice_uuidEvents
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      example: '2026-05-16 11:15:30'
                    type:
                      type: string
                      example: fixed
                    title:
                      type: string
                      example: 'Naprawiono usterkę'
                    description:
                      type: string
                      example: 'Zgłoszenie zostało pomyślnie zamknięte i oznaczone jako rozwiązane.'
                    user:
                      type: string
                      example: Serwisant
                example:
                  -
                    date: '2026-05-16 11:15:30'
                    type: fixed
                    title: 'Naprawiono usterkę'
                    description: 'Zgłoszenie zostało pomyślnie zamknięte i oznaczone jako rozwiązane.'
                    user: Serwisant
                  -
                    date: '2026-05-16 11:06:47'
                    type: install
                    title: 'Instalacja urządzenia'
                    description: 'Pierwszy montaż i uruchomienie w systemie.'
                    user: System
                  -
                    date: '2026-05-16 11:06:47'
                    type: fault
                    title: 'Zgłoszono usterkę'
                    description: 'Nadmierne wibracje'
                    user: 'Anna Nowak'
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  '/api/devices/{device_uuid}/faults':
    post:
      summary: ''
      operationId: postApiDevicesDevice_uuidFaults
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                description:
                  type: string
                  description: ''
                  example: 'Eius et animi quos velit et.'
                  nullable: true
                reported_by:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: v
                  nullable: true
                contact:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: d
                  nullable: true
              required:
                - title
      security: []
    get:
      summary: ''
      operationId: getApiDevicesDevice_uuidFaults
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
...
  /api/user/password:
    put:
      summary: ''
      operationId: putApiUserPassword
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: ''
                  example: architecto
                password:
                  type: string
                  description: ''
                  example: architecto
              required:
                - current_password
                - password
      security: []
...
  '/api/faults/{fault_id}':
    get:
      summary: ''
      operationId: getApiFaultsFault_id
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    patch:
      summary: ''
      operationId: patchApiFaultsFault_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: resolved
                  enum:
                    - pending
                    - in_progress
                    - resolved
              required:
                - status
      security: []
    parameters:
      -
        in: path
        name: fault_id
        description: 'The ID of the fault.'
        example: 3
        required: true
        schema:
          type: integer
...
  /api/devices:
    get:
      summary: ''
      operationId: getApiDevices
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiDevices
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: 'Must be a valid UUID.'
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: true
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: g
                type:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: z
                model:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: m
                  nullable: true
                brand:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: i
                  nullable: true
                serial_number:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'y'
                  nullable: true
                location:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: v
                installation_date:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2026-05-22T20:53:28'
                  nullable: true
                notes:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                instruction_url:
                  type: string
                  description: 'Must be a valid URL. Must not be greater than 2048 characters.'
                  example: 'http://bailey.com/'
                  nullable: true
              required:
                - name
                - type
                - location
      security: []
  '/api/devices/{uuid}':
    put:
      summary: ''
      operationId: putApiDevicesUuid
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                type:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                model:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: g
                  nullable: true
                brand:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: z
                  nullable: true
                serial_number:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: m
                  nullable: true
                location:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: i
                installation_date:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2026-05-22T20:53:28'
                  nullable: true
                notes:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                instruction_url:
                  type: string
                  description: 'Must be a valid URL. Must not be greater than 2048 characters.'
                  example: 'http://bailey.com/'
                  nullable: true
      security: []
    delete:
      summary: ''
      operationId: deleteApiDevicesUuid
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  /api/devices/generate-qr:
    post:
      summary: ''
      operationId: postApiDevicesGenerateQr
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: 'Must be a valid UUID.'
                  example: 6ff8f7f6-1eb3-3525-be4a-3932c805afed
                  nullable: true
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: g
                type:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: z
                model:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: m
                  nullable: true
                brand:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: i
                  nullable: true
                serial_number:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'y'
                  nullable: true
                location:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: v
                installation_date:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2026-05-22T20:53:28'
                  nullable: true
                notes:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                instruction_url:
                  type: string
                  description: 'Must be a valid URL. Must not be greater than 2048 characters.'
                  example: 'http://bailey.com/'
                  nullable: true
              required:
                - name
                - type
                - location
      security: []
  '/api/devices/{device_uuid}/show-qr':
    get:
      summary: ''
      operationId: getApiDevicesDevice_uuidShowQr
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  '/api/devices/{device_uuid}/instruction':
    patch:
      summary: ''
      operationId: patchApiDevicesDevice_uuidInstruction
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                instruction_url:
                  type: string
                  description: 'Must be a valid URL. Must not be greater than 2048 characters.'
                  example: 'http://www.bailey.biz/quos-velit-et-fugiat-sunt-nihil-accusantium-harum.html'
                  nullable: true
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  '/api/devices/{device_uuid}/assign':
    post:
      summary: ''
      operationId: postApiDevicesDevice_uuidAssign
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  description: 'The <code>id</code> of an existing record in the users table.'
                  example: 16
              required:
                - user_id
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  '/api/devices/{device_uuid}/assign/{user_id}':
    delete:
      summary: ''
      operationId: deleteApiDevicesDevice_uuidAssignUser_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/devices/{device_uuid}/users':
    get:
      summary: ''
      operationId: getApiDevicesDevice_uuidUsers
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: device_uuid
        description: ''
        example: f92b0ee1-d31c-3b96-bd76-0f1840d8d289
        required: true
        schema:
          type: string
  '/api/users/{user_id}/devices':
    get:
      summary: ''
      operationId: getApiUsersUser_idDevices
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/users/{id}':
    put:
      summary: ''
      operationId: putApiUsersId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: zbailey@example.net
                password:
                  type: string
                  description: ''
                  example: null
                  nullable: true
                is_admin:
                  type: boolean
                  description: ''
                  example: false
                is_installer:
                  type: boolean
                  description: ''
                  example: true
                is_service:
                  type: boolean
                  description: ''
                  example: true
      security: []
    delete:
      summary: ''
      operationId: deleteApiUsersId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...